requirements.txt
# Core dependencies
json5>=0.9.0
pathlib>=1.0.1
python-dotenv>=1.0.0
requests>=2.25.0

# LLM dependencies
openai>=1.51.0
anthropic>=0.3.0
google-genai>=0.2.0

# Vector search and ML dependencies
faiss-cpu>=1.7.4
numpy>=1.21.0
tiktoken>=0.4.0
